modular dash refactor
